+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
+### Added
+
+- Added TTFB metrics for `HeyGenVideoService` and `TavusVideoService`.
+
 ### Other
 
 - Updated `15-switch-voices.py` and `15a-switch-languages.py` examples to show